// src/types/table.types.ts
// src/types/table.types.ts

export interface DataGridRow {
  id: number;
  shortName: string;
  fullName: string;
  price: number;
  quantity: number;
  date: Date;
  comment1: string;
  comment2: string;
  comment3: string;
  comment4: string;
  comment5: string;
  comment6: string;
  comment7: string;
  comment8: string;
  comment9: string;
  comment10: string;
}